backend/requirements.txt
alembic==1.4.2
astroid==2.4.2
backcall==0.2.0
click==7.1.2
decorator==4.4.2
fastapi==0.58.1
h11==0.9.0
httptools==0.1.1
ipython-genutils==0.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.2
pickleshare==0.7.5
ptyprocess==0.6.0
pydantic==1.5.1
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.4
toml==0.10.1
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
